pyproject.toml
[project]
name = "rl-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gymnasium[classic-control]>=1.2.3",
    "ipykernel>=7.1.0",
    "numpy>=2.4.1",
    "torch>=2.10.0",
    "hydra-core>=1.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rl_training"]
